Conflicts:
	ProyectG10/src/main/java/com/mycompany/proyectg10/Operador.java

+++ ProyectG10/src/main/java/com/mycompany/proyectg10/Cliente.java
@@ -164,7 +164,7 @@ public class Cliente extends Usuario {
                     int codigo = rd.nextInt(9000)+1000;
                     Revision rev = new Revision(codigo,getNumCedula(),vehiculo.getNumPlaca(),fechaRevision);
                     ManejoArchivos.EscribirArchivo("AgendaRevisiones.txt", rev.escribirRevision());
-                    int totalAPagar = 0;
+                    double totalAPagar = 0;
                     for (Multa mult : multasCliente){
                         totalAPagar += rev.valorRevision(mult);
                     }
